Support FTP folder destination
In the ftp-config file, you can precise in which folder should be uploaded the app.
It is useful for example when your ftp root folder is not the web root folder.

Note : you can still only access the app from the web domain root. If you have decided to upload the app in the "www/my-app/" folder, you will not be able to access it through "http://my-domain.com/my-app". You will either have to upload it to the folder "www/" so you can access it through "http://my-domain.com/" or you can create a subdomain pointing to "www/my-app/" folder so you can access the app through "http://my-subdomain.my-domain.com/"